feat: use adapter service to access api

+++ internal/core/domain/entity/cpu.go
@@ -1,10 +1,10 @@
 package entity
 
 type Cpu struct {
-	Value int64
+	Value int
 	Unit  string
 }
 
-func NewCpu(value int64, unit string) Cpu {
+func NewCpu(value int, unit string) Cpu {
 	return Cpu{Value: value, Unit: unit}
 }
