+++ internal/core/domain/entity/cpu_test.go
@@ -9,6 +9,6 @@ import (
 func TestNewCpu(t *testing.T) {
 	cpu := NewCpu(1, "unit")
 
-	assert.Equal(t, int64(1), cpu.Value)
+	assert.Equal(t, 1, cpu.Value)
 	assert.Equal(t, "unit", cpu.Unit)
 }
