+++ internal/core/domain/entity/network_speed_test.go
@@ -9,6 +9,6 @@ import (
 func TestNewNetworkSpeed(t *testing.T) {
 	networkSpeed := NewNetworkSpeed(1, "unit")
 
-	assert.Equal(t, int64(1), networkSpeed.Value)
+	assert.Equal(t, 1, networkSpeed.Value)
 	assert.Equal(t, "unit", networkSpeed.Unit)
 }
