+++ internal/core/domain/entity/sticky_session_test.go
@@ -10,5 +10,5 @@ func TestNewStickySession(t *testing.T) {
 	got := NewStickySession(true, 5)
 
 	assert.True(t, got.Enabled)
-	assert.Equal(t, int64(5), got.MaxLifeTime)
+	assert.Equal(t, 5, got.MaxLifeTime)
 }
